feature/Check-PSVersion (#64)
* Bugfix: Get-EsetLink regex failure

Failed to take dashes into account when matching for org name in Get-EsetLink (line 593). Not an issue for 99.99% of clients in Kaseya but broke on "test-org-arjun". Changed regex to "[\w-]+$" to accomodate.

Also alphabetized the Readme again, just because.

* feature/Check-PSVersion

Added function Check-PSVersion to check if installed PowerShell is less than 3. If so, return True (optional: e-mail managed services to schedule an update).

* feature/Check-PSVersion

Added function Check-PSVersion to check if installed PowerShell is less than 3. If so, return True (optional: e-mail managed services to schedule an update).

* Combined and fixed Readme

There was a conflict with the Readme, and the formatting was all weird (like line breaks not actually creating a new line). Fixed.

* Weird typos

There was an errant close-bracket; I think it was added by the conflict manager? Also had missed a(n invisible) linebreak character last time I fixed the Readme. Cleaned up.