ReWrote Get-DiskChanges (#78)
* Updated Get-DiskStatistics

Function Get-DiskStatistics will now only return the fixed disks on the computer.  It will skip removable drives and cd drives.

* BugFixes in Get-DiskChanges

Changed line to out-null when creating a new folder.  No needed.  Cleaned up the help information in Get-DiskChanges and Get-DriveStatistics.  Added inline comment for comparison block in case someone wants to copy the code to make another comparison.

* Created Get-VSSStatistics

New function that returns the state of VSS snapshots upsed.

* Updated Get-VSSStatistics & Get-DiskChanges

Finished writing Get-VSSStatistics and all error handling as well as updating Get-DiskChanges to also account for Get-VSSStatistics.  Tested to make sure that files are working correctly in both admin and non admin powershell sessions.  Created all help files and updated readme.md.

* Found bug in Get-DiskChanges for VSS

Found that it will not get the diskchanges unless there was an exact match in shadowstorage volumes.  Modified script so that it will still run if there is a mismatch of volumes in shadowstorage volumes.  It will write all zeros to the columns if that is the case.

* ReWrote Get-DiskChanges

Found that Get-DiskChanges didn't work as intended in the field.  Rewrote while tesing on a production server. Verified that it is now working.