Updated Files (#81)
* Updated Get-DiskStatistics

Function Get-DiskStatistics will now only return the fixed disks on the computer.  It will skip removable drives and cd drives.

* Bugfix: Get-EsetLink regex failure

Failed to take dashes into account when matching for org name in Get-EsetLink (line 593). Not an issue for 99.99% of clients in Kaseya but broke on "test-org-arjun". Changed regex to "[\w-]+$" to accomodate.

Also alphabetized the Readme again, just because.

* BugFixes in Get-DiskChanges

Changed line to out-null when creating a new folder.  No needed.  Cleaned up the help information in Get-DiskChanges and Get-DriveStatistics.  Added inline comment for comparison block in case someone wants to copy the code to make another comparison.

* feature/Check-PSVersion

Added function Check-PSVersion to check if installed PowerShell is less than 3. If so, return True (optional: e-mail managed services to schedule an update).

* feature/Check-PSVersion

Added function Check-PSVersion to check if installed PowerShell is less than 3. If so, return True (optional: e-mail managed services to schedule an update).

* Combined and fixed Readme

There was a conflict with the Readme, and the formatting was all weird (like line breaks not actually creating a new line). Fixed.

* Weird typos

There was an errant close-bracket; I think it was added by the conflict manager? Also had missed a(n invisible) linebreak character last time I fixed the Readme. Cleaned up.

* Created Get-VSSStatistics

New function that returns the state of VSS snapshots upsed.

* Updated Get-VSSStatistics & Get-DiskChanges

Finished writing Get-VSSStatistics and all error handling as well as updating Get-DiskChanges to also account for Get-VSSStatistics.  Tested to make sure that files are working correctly in both admin and non admin powershell sessions.  Created all help files and updated readme.md.

* Found bug in Get-DiskChanges for VSS

Found that it will not get the diskchanges unless there was an exact match in shadowstorage volumes.  Modified script so that it will still run if there is a mismatch of volumes in shadowstorage volumes.  It will write all zeros to the columns if that is the case.

* ReWrote Get-DiskChanges

Found that Get-DiskChanges didn't work as intended in the field.  Rewrote while tesing on a production server. Verified that it is now working.

* Created Deploy-DiskChanges

New cmdlet that will schedulde get-diskchanges on a machine to run every 10 minutes.  Still need to figure out the user account that it will run as.

* TEMP

* Merge and minor update

Merged IntegratedITSolutionsInc/master into master. Also made minor update to Get-EsetLink (changed alias to Get-EsetAgent rather than Get-Eset, which I plan to use as the name of another function).

* Frequent Script Examples

Created two template-style example scripts for frequent reference: one for error logging and one for joining two lists.

* Created Send-PatchEmail 

This function will figure out if it is an appropriate day to send out a patching email.  need to configure the if statements that connect to the IITS exchange server to send out the email to the patching distribution list.

* Error handling for Get-EsetLink

Added error handling for Get-EsetLink, including a new $Logging parameter. Also made minor revision to Error Logging example script.

* Minor updates to Get-EsetLink

Minor revisions after Darren's review. Changed $machName from a parameter to a self-determined variable.

* Minor cleanup update

Slight update to help text for Check-PSVersion and Get-EsetLink. Removed a few extraneous newlines between functions to better standardize spacing (not that it ultimately matters). Changed all instances of "return" to "echo" in Check-PSVersion for better overall performance. Decided *against* adding logging for Check-PSVersion; it's so simple a function that adding error logging feels like gross overengineering.

* Finished Send-PatchEmail

Added in the email pieces as well as connecting to the exchange server.  This can only be run on the KUTIL server.