Updated Send-PatchEmail (#88)
* Had to modify the error file name

Needed to change it to patchemail.txt

* Email formatting

Send-Patchemail had to be reformatted to account for the extra tab spaces that powershell ISE automatically enters.

* Updated Send-PatchEmail

Updated function to send email to iits_Patching_clients as bcc and not to.  This way they can't reply all and see the rest of the clients.   Good catch Geo R.