New/Install-Eset
Added Check-EsetAgent to README list. Recreated shell for Check-EsetAgent function (it was easier to overwrite than to fix a big merge), then wrote the function.

Made minor changes to Error Logging exmaple. Started work on Install-Eset. Added Install-Eset to README. Created a ton of (empty) helper programs for use with Install-Eset; still need to add them to README but not until I verify thier use.

Lots of development on Install-Eset but not yet complete. Error- and exception-handling is an important but constant speedbump.