Added Run-NetworkDetective (#117)
* Created program that will compare exes

This will alert if there is a new executable or a removed executable on the server.

* Need to fix the deletion code at the end

Takes 7 minutes to run.

* Rewrote file to use compare-object and not a nuch of foreach statements

Function ran in 10 seconds per fun instead of 7 minutes like before.  Fixed deletion script.

* Created Get-ActivationStatus

This cmdlet will find if a machine is not licensed.  It will then email msalarm if a machine is found to be unlicensed.

* Created Run-NetworkDetective

Created function that will run 3 network detective scans.  It will then email the results into msalarm.