amendGet-Projection (#122)
* Added/modified function (Delete-IITStasks/Get-Projection resp.)

Added function to delete IITS tasks and modified function get-projection

* Added function checkIITStasks

The function will check the last 2 consecutive runs of custom IITS tasks and will send an e-mail if 2 consecutive runs fail

* amended the Get-Projection Script

Changed the verbiage of the Get-Projection script. To include a valid subject and some meaningful lines.

* amendedGet-Projection

There was a logical error in the script. Corrected the Logic